% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{check_aliases}
\alias{check_aliases}
\title{Check font aliases for missing families and faces}
\usage{
check_aliases(aliases)
}
\arguments{
\item{aliases}{A list of families of faces. Typically created with
\code{fontquiver::font_families()}.}
}
\description{
Check font aliases for missing families and faces
}

