% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_svgz}
\alias{create_svgz}
\title{Convert an svg file to svgz, overwriting the old file}
\usage{
create_svgz(file)
}
\arguments{
\item{file}{the path to the file to convert}
}
\description{
Convert an svg file to svgz, overwriting the old file
}
\keyword{internal}
