% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SVG.R
\name{devSVG}
\alias{devSVG}
\title{A SVG Graphics Driver}
\usage{
devSVG(file = "Rplots.svg", width = 10, height = 8, bg = "white",
  fg = "black", onefile = TRUE, xmlHeader = TRUE, useNS = TRUE)
}
\arguments{
\item{file}{the file where output will appear.}

\item{width}{The width of the plot in inches.}

\item{height}{the height of the plot in inches.}

\item{bg}{the background color for the plot.}

\item{fg}{the foreground color for the plot.}

\item{onefile}{merge plot calls into onefile or separate them to separate pages.}

\item{xmlHeader}{Print XML header or not.}

\item{useNS}{Include svg namespace? Not including it makes extracting
elements with xpath much easier.}
}
\description{
This function produces graphics suitable the current w3 svg XML standard.
It currently does not have any font metric information, so the use of
\code{\link{plotmath}} is not supported. The driver output is currently NOT
specifying a DOCTYPE DTD
}
\examples{
devSVG()
plot(1:11,(-5:5)^2, type='b', main="Simple Example")
dev.off()
}
\author{
This driver was written by T Jake Luciani
  \email{jakeluciani@yahoo.com} 2012: updated by Matthieu Decorde
  \email{matthieu.decorde@ens-lyon.fr}
}
\references{
\emph{W3C Scalable Vector Graphics (SVG)}:
  \url{http://www.w3.org/Graphics/SVG/Overview.htm8}
}
\seealso{
\code{\link{pictex}}, \code{\link{postscript}}, \code{\link{Devices}}
}
\keyword{device}

