% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inlineSVG.R
\name{editSVG}
\alias{editSVG}
\title{Run plotting code and open svg in OS/system default svg viewer or editor.}
\usage{
editSVG(code, ..., width = NA, height = NA, font_aliases = NULL,
  font_extra = NULL)
}
\arguments{
\item{code}{Plotting code to execute.}

\item{...}{Other arguments passed on to \code{\link{svglite}}.}

\item{width}{Height and width in inches.}

\item{height}{Height and width in inches.}

\item{font_aliases}{List of font families to be aliased. By
default, \code{sans} is aliased to "Arial", \code{serif} to
"Times" or "Times New Roman" (the latter on Windows), and
\code{mono} to "courier".}

\item{font_extra}{List of fonts or collection of fonts. These are
used by svglite to compute text metrics in a reproducible way.}
}
\description{
This is useful primarily for testing or post-processing the SVG.
}
\examples{
if (interactive()) {
  editSVG(plot(1:10))
  editSVG(contour(volcano))
}
}

