% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{register_variant}
\alias{register_font}
\alias{font_feature}
\alias{fonts_as_import}
\alias{add_fonts}
\alias{require_font}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{systemfonts}{\code{\link[systemfonts]{add_fonts}}, \code{\link[systemfonts]{font_feature}}, \code{\link[systemfonts]{fonts_as_import}}, \code{\link[systemfonts]{register_font}}, \code{\link[systemfonts]{register_variant}}, \code{\link[systemfonts]{require_font}}}
}}

