% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVG.R
\name{stringSVG}
\alias{stringSVG}
\title{Extract the content of SVG as a string}
\usage{
stringSVG()
}
\value{
A character string representing the last plot created by \code{svglite(":string:")}.
}
\description{
This function is used to extract the content of SVG produced by
the most recent call of \code{svglite(":string:")},
which directly writes the SVG to a string stream
and does not generate any temporary files.
}
\examples{
svglite(":string:")
plot(rnorm(5), rnorm(5))
dev.off()
svgstr = stringSVG()
cat(svgstr)
}

