% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svglite-package.R
\docType{package}
\name{svglite-package}
\alias{svglite-package}
\alias{_PACKAGE}
\title{svglite: An 'SVG' Graphics Device}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A graphics device for R that produces 'Scalable Vector Graphics'. 'svglite' is a fork of the older 'RSvgDevice' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://svglite.r-lib.org}
  \item \url{https://github.com/r-lib/svglite}
  \item Report bugs at \url{https://github.com/r-lib/svglite/issues}
}

}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomas.pedersen@posit.co} (\href{https://orcid.org/0000-0002-5147-4711}{ORCID})

Authors:
\itemize{
  \item Hadley Wickham \email{hadley@posit.co}
  \item Lionel Henry \email{lionel@posit.co}
  \item T Jake Luciani \email{jake@apache.org}
  \item Matthieu Decorde \email{matthieu.decorde@ens-lyon.fr}
  \item Vaudor Lise \email{lise.vaudor@ens-lyon.fr}
}

Other contributors:
\itemize{
  \item Tony Plate (Early line dashing code) [contributor]
  \item David Gohel (Line dashing code and early raster code) [contributor]
  \item Yixuan Qiu (Improved styles; polypath implementation) [contributor]
  \item Håkon Malmedal (Opacity code) [contributor]
  \item Posit, PBC [copyright holder, funder]
}

}
\keyword{internal}
