% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svghtmldev.R
\name{svglitedev}
\alias{svglitedev}
\title{A fixed-size SVG graphics device for iterative plot development}
\usage{
svglitedev(width = 10, height = 8, bg = "white", pointsize = 12)
}
\arguments{
\item{height, width}{Height and width in inches.}

\item{bg}{Default background color for the plot (defaults to "white").}

\item{pointsize}{default point size.}

\item{file}{the file where output will appear.}
}
\description{
Set \code{svglitedev()} as your graphics device at the start of a session
and your plotting commands will appear in the R Studio viewer or browser
window.  Then use svglite(), knitr, or another tool to produce the final file.
}
\examples{
svglitedev()
plot(1:10)
hist(rnorm(100))

}

