% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inlineSVG.R
\name{editSVG}
\alias{editSVG}
\title{Run plotting code and open svg in OS/system default svg viewer or editor.}
\usage{
editSVG(code, ..., width = NA, height = NA, system_fonts = list(),
  user_fonts = list())
}
\arguments{
\item{code}{Plotting code to execute.}

\item{...}{Other arguments passed on to \code{\link{svglite}}.}

\item{width}{Height and width in inches.}

\item{height}{Height and width in inches.}

\item{system_fonts}{Named list of font names to be aliased with
fonts installed on your system. If unspecified, the R default
families \code{sans}, \code{serif}, \code{mono} and \code{symbol}
are aliased to the family returned by
\code{\link[gdtools]{match_family}()}.}

\item{user_fonts}{Named list of fonts to be aliased with specific
font files. These font files can be fonts from the fontquiver
package or they}
}
\description{
This is useful primarily for testing or post-processing the SVG.
}
\examples{
if (interactive()) {
  editSVG(plot(1:10))
  editSVG(contour(volcano))
}
}

