% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inlineSVG.R
\name{htmlSVG}
\alias{htmlSVG}
\title{Run plotting code and view svg in RStudio Viewer or web broswer.}
\usage{
htmlSVG(code, ..., system_fonts = list(), user_fonts = list())
}
\arguments{
\item{code}{Plotting code to execute.}

\item{...}{Other arguments passed on to \code{\link{svglite}}.}

\item{system_fonts}{Named list of font names to be aliased with
fonts installed on your system. If unspecified, the R default
families \code{sans}, \code{serif}, \code{mono} and \code{symbol}
are aliased to the family returned by
\code{\link[gdtools]{match_family}()}.}

\item{user_fonts}{Named list of fonts to be aliased with specific
font files. The aliases can be fonts from the fontquiver package,
strings pointing to a font file, or a list containing \code{name}
and \code{file} elements with \code{name} indicating the font
alias in the SVG output.}
}
\description{
This is useful primarily for testing. Requires the \code{htmltools}
package.
}
\examples{
if (require("htmltools")) {
  htmlSVG(plot(1:10))
  htmlSVG(hist(rnorm(100)))
}
}

