% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inlineSVG.R
\name{htmlSVG}
\alias{htmlSVG}
\title{Run plotting code and view svg in RStudio Viewer or web broswer.}
\usage{
htmlSVG(code, ...)
}
\arguments{
\item{code}{Plotting code to execute.}

\item{...}{Other arguments passed on to \code{\link{svglite}}.}
}
\description{
This is useful primarily for testing. Requires the \code{htmltools}
package.
}
\examples{
if (require("htmltools")) {
  htmlSVG(plot(1:10))
  htmlSVG(hist(rnorm(100)))
}
}
