% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inlineSVG.R
\name{stringSVG}
\alias{stringSVG}
\title{Run plotting code and return svg as string}
\usage{
stringSVG(code, ...)
}
\arguments{
\item{code}{Plotting code to execute.}

\item{...}{Other arguments passed on to \code{\link{svglite}}.}
}
\description{
This is useful primarily for testing but can be used as an
alternative to \code{\link{svgstring}()}.
}
\examples{
stringSVG(plot(1:10))
}
