% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xmlSVG.R
\name{xmlSVG}
\alias{xmlSVG}
\title{Run plotting code and return svg}
\usage{
xmlSVG(code)
}
\arguments{
\item{code}{Plotting code to execute.}
}
\value{
A \code{xml2::xml_document} object.
}
\description{
This is useful primarily for testing. Requires the \code{xml2} package.
}
\examples{
if (require("xml2")) {
  x <- xmlSVG(plot(1, axes = FALSE))
  x
  xml_find_all(x, ".//text")
}
}

